Split DocumentRenderPane into scrollable- and non-scrollable variant
(cherry picked from commit 44dd27edfd967745c548f7ae1fd2047e812cdc22)
